% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylegend.R
\name{asCol}
\alias{asCol}
\title{Converts categorical values to colors. Wrapper for setFactorGraphics.}
\usage{
asCol(x, export = FALSE, ...)
}
\arguments{
\item{x}{A vector of numerical, character, boolean or factor values, 
representing categories (non-continuous, with a limited number 
of unique values).}

\item{export}{Single logical value. If \code{TRUE}, the result of 
\code{\link[easylegend]{setFactorGraphics}} will be saved into 
the output attribute \code{fg}.}

\item{\dots}{Additional parameters passed to \code{\link[easylegend]{setFactorGraphics}}.}
}
\value{
Returns a list of colors, with \code{\link[base]{attr}}ibutes 
 \code{fg} set to output of \code{\link[easylegend]{setFactorGraphics}}.
}
\description{
Converts categorical values to colors. Wrapper for setFactorGraphics.
}
\examples{
library( "easylegend" ) 


# Example: x numerical values, points =======================
x <- sample( x = round(rnorm(10),2), size = 20, replace = TRUE )

asCol( x ) 
asCol( x, export = TRUE ) 
}

