% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylegend.R
\name{asPch}
\alias{asPch}
\title{Converts categorical values to symbols. Wrapper for setFactorGraphics.}
\usage{
asPch(x, export = FALSE, ...)
}
\arguments{
\item{x}{A vector of numerical, character, boolean or factor values, 
representing categories (non-continuous, with a limited number 
of unique values).}

\item{export}{Single logical value. If \code{TRUE}, the result of 
\code{\link[easylegend]{setFactorGraphics}} will be saved into 
the output attribute \code{fg}.}

\item{\dots}{Additional parameters passed to \code{\link[easylegend]{setFactorGraphics}}.}
}
\value{
Returns a list of symbols.
}
\description{
Converts categorical values to symbols. Wrapper for setFactorGraphics.
}
\examples{
library( "easylegend" ) 


# Example: x numerical values, points =======================
x <- sample( x = round(rnorm(10),2), size = 20, replace = TRUE )

asPch( x ) 
asPch( x, export = TRUE ) 
}

