% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylegend.R
\name{matrix2image}
\alias{matrix2image}
\title{Arrange a matrix for display with image() that respects original rows and columns}
\usage{
matrix2image(x)
}
\arguments{
\item{x}{An \code{\link[base]{matrix}}}
}
\value{
Return a transposed and column-flipped matrix
}
\description{
Arrange a matrix for display with image() that respects original rows and columns
}

