% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylegend.R
\name{plotAnywhere}
\alias{plotAnywhere}
\title{Plot (a legend) anywhere in the figure (margins included)}
\usage{
plotAnywhere(expr = NULL, mar = rep(0.1, 4), .clip = TRUE, ...)
}
\arguments{
\item{expr}{An \code{\link[base]{expression}}, presumably some plot-overlay.}

\item{mar}{See \code{\link[graphics]{par}}.}

\item{.clip}{Single logical value. If \code{TRUE} (default), the 
\code{\link[graphics]{clip}}pping region is preserved too.}

\item{\dots}{Additional parameters passed to \code{\link[graphics]{par}}.}
}
\value{
Invisibly returns the original \code{\link[graphics]{par}} 
 settings
}
\description{
Plot (a legend) anywhere in the figure (margins included). Creates 
 a temporary figure-overlay with extra small margins (see 
 \code{mar}) and with figure extent of 0 to 1 in the x and y 
 axis, so that some plot (a legend) can be overlain on top 
 of an existing figure.
}

