% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylegend.R
\name{setColourRampScale}
\alias{setColourRampScale}
\alias{setColourRampScale.RasterLayer}
\alias{setColourRampScale.default}
\alias{setColourRampScale.matrix}
\title{Prepare colour scale (ramp) and legend (fill) from continuous-numeric data.}
\usage{
setColourRampScale(x, ...)

\method{setColourRampScale}{default}(x, fill = NULL, breaks = NULL,
  int = 4L, naCol = "lightgray", naLeg = "na", brackets = c("", " ",
  ""), labels = NULL, right = FALSE, include.lowest = TRUE, digits = 3L,
  nsmall = 3L, decreasing = TRUE, y.intersp = 1.5, nBreaks = 5L,
  alpha = FALSE, hideMinMax = c(FALSE, FALSE), ...)

\method{setColourRampScale}{matrix}(x, ...)

\method{setColourRampScale}{RasterLayer}(x, ...)
}
\arguments{
\item{x}{A vector of numerical values (continous).}

\item{fill}{Logical value or vector of character strings representing 
fill-colours. If \code{TRUE} or vector of colours, a 
continuous fill-colour scale is defined for 'x'. There 
must be as many fill-colours as \code{breaks}.}

\item{breaks}{See \code{\link[base]{cut}}. There must be as many breaks 
as colours (i.e. \code{length(fill) == length(breaks)}; 
Notice that this is different from 
\code{\link{setColourScale}}). If \code{NULL}, 
\code{nBreaks} breaks will be generated internally (see 
below). \code{breaks} should be ordered (no \code{NA}), 
with no replica, and consistent with the argument 
\code{decreasing}. See \code{naCol} and \code{naLeg} 
regarding the handling of \code{NA}-values.}

\item{int}{Single integer value. Number of colour intervals to be 
defined in between the main fill-colours. \code{int} must 
be \code{> 1}. Intermediate colours are generated with 
\code{\link[grDevices]{colorRampPalette}}.}

\item{naCol}{Single character string. Colour for \code{NA} (missing) 
values. Notice that in the case of an 
\code{\link[graphics]{image}} plot this parameter is 
silently be ignored and replaced by \code{NA} as 
the function displays missing values as transparent pixels.
In the case of \code{raster::plot}, One should make sure 
that the value of \code{naCol}-argument in 
\code{setColourRampScale} is coherent with the value of 
\code{naCol}-argument in \code{raster::plot}.}

\item{naLeg}{Single character string. Legend (label) for \code{NA} 
(missing) values.}

\item{brackets}{Vector of 3 characters. Used to generate customise 
legend-labels, if \code{labels} is \code{NULL}.}

\item{labels}{See \code{\link[base]{cut}}.}

\item{right}{See \code{\link[base]{cut}}. This parameter is important 
as it will determine how \code{x}-values falling on a break 
will be classified.}

\item{include.lowest}{See \code{\link[base]{cut}}.}

\item{digits}{See \code{\link[base]{format}}. Number formatting in the 
legend.}

\item{nsmall}{See \code{\link[base]{format}}. Number formatting in the 
legend.}

\item{decreasing}{Logical value. If \code{TRUE} (default), the legend is 
ordered (sorted) in decreasing order (i.e. low values on 
the bottom of the colour scale and high values on the top 
of the colour scale).}

\item{y.intersp}{Single numerical value. Character inter-spacing factor 
for vertical (y) spacing of the colour legend. Passed to 
\code{\link[graphics]{legend}}.}

\item{nBreaks}{Single integer value. Number of breaks that must be 
generated internally when \code{breaks} is not specified 
(\code{NULL}). See argument \code{breaks} for details 
on how \code{nBreaks} and \code{fill} are related.}

\item{alpha}{Single logical value. Passed internally to 
\code{\link[grDevices]{colorRampPalette}}}

\item{hideMinMax}{Vector of two logical value. If \code{TRUE}, the min or 
the max values in breaks (or both) will be replaced by 
\code{> x} or \code{< y} where \code{x} is the 2nd highest 
value in \code{breaks} and \code{y} is the 2nd smallest 
value in \code{breaks}. \code{hideMinMax[1L]} is for the 
min value and \code{hideMinMax[2L]} is for the max-value.}

\item{\dots}{Additional parameters passed to specific methods.}
}
\value{
Returns a list of 2 \code{\link[base]{function}}: 
 \code{$fill()}, a function that converts \code{x}-like values 
 into (fill)colours; \code{$legend()}, a function to draw a 
 legend on a plot with the correct (fill)colours and legend. 
 That function accepts \code{...} arguments, passed to the 
 original \code{\link[graphics]{legend}} function.
}
\description{
Prepare colour scale (ramp) and legend (fill) from 
 continuous-numeric data. \code{setColourScale} is a 
 partial rewriting of \code{\link{setColorScale}}.
}
\examples{
# See ?setColourScale for a series of examples
}
\seealso{
\code{\link{setColourScale}}, 
 \code{\link{setFactorGraphics}} and 
 \code{\link[grDevices]{colorRampPalette}}.
}

